Add files via upload
The Plotly Dash app primarily consists of two plots, namely a pie chart displaying the success-failure distribution with respect to launch sites and a scatter plot between launch class and payload mass visualizing their relation and distribution. 
The pie chart can be interacted using a dropdown menu consisting of all the launch sites, letting us gain insights for each launch site separately and all sites altogether.
A range slider is added for the scatter plot which has the range for the payload mass which lets us interact with the chart in a way which lets us visualize the success-failure distribution depending on the payload mass and the launch site as well. Alongside that, the markers are colour-coded with respect to the booster version used, letting the user visualize how success-failure distribution is related to the payload and various booster versions as well.